requirements.txt
fastapi==0.103.1
uvicorn==0.22.0
pandas==2.2.1
openpyxl==3.1.2
google-cloud-storage==2.14.0
python-multipart==0.0.6
